+++ docs/source/changelog.rst
@@ -21,6 +21,11 @@ Changelog
 * :support:`3` Co-integration approach documentation.
 * :feature:`4` Landmark paper: PCA Approach (Avellaneda and Lee, 2010)
 * :support:`4` Documentation for PCA approach.
+* :feature:`5` ML Based Pairs Selection (Horta, 2020) and Data Importer added.
+* :support:`5` ML Based Pairs Selection and Data Importer documentation.
+* :feature:`8` Copula strategy added (Liew et al. 2013): Log price (or equivalently, cumulative log returns) based copula strategy.
+* :feature:`8` Copulas supported: Gumbel, Frank, Clayton, Joe, N13, N14, Gaussian, Student(Student-t).
+* :support:`8` Copula strategy documentation (Liew et al. 2013) for log price based copula strategy.
 * :feature:`14` Landmark paper: The Distance Approach (Gatev et al. 2006).
 * :support:`14` Distance approach documentation.
 * :support:`14` Added a number of new tools to improve our deployment and align us with best practices. They include: CircleCI, VersionBump, Update Issue Templates, ChangeLog, Logo, Favicon.
@@ -29,8 +34,9 @@ Changelog
 * :feature:`16` OU-model to the Optimal Mean Reversion module added.
 * :support:`16` OU-model documentation.
 * :support:`17` Added Licence, ReadMe, and RoadMap
+* :feature:`19` Minimum profit optimization module added (Lin et al. 2006, Puspaningrum et al. 2010): Finding optimal pre-set boundaries for cointegrated pairs trading strategy.
+* :feature:`19` Cointegrated time series simulation module added (Lin et al. 2006): Simulate cointegrated series that follows AR(1) dynamics.
+* :support:`19` Minimum profit optimization documentation for cointegrated pairs trading strategy.
+* :support:`19` Cointegrated time series simulation documentation.
 * :support:`20` Added API Key Validation.
 * :support:`20` Add install documentation and test on OS/Ubuntu/Windows.
-* :feature:`8` Copula strategy added (Liew et al. 2013): Log price (or equivalently, cumulative log returns) based copula strategy.
-* :feature:`8` Copulas supported: Gumbel, Frank, Clayton, Joe, N13, N14, Gaussian, Student(Student-t).
-* :support:`8` Copula strategy documentation (Liew et al. 2013) for log price based copula strategy.